fixed a crash bug when calcualte a normalized serial coordinate, the value is of integer limit. Changed to double.

+++ src/ck-ldb/OrbLB.h
@@ -39,7 +39,7 @@ private:
     int v[3];
     double load;
     int  refno;
-    int  tv;
+    double  tv;
     Partition * partition;
   };
